# spring/classes/openapi/api-spec.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Abacus WEB Sample",
    "description": "Abacus WEB Sample Project",
    "license": {
      "name": "Apache 2.0",
      "url": "http://abacus.co.kr"
    },
    "version": "v1.0"
  },
  "servers": [
    {
      "url": "http://192.168.2.142:8080/api/",
      "description": "dev"
    }
  ],
...
  "paths": {
...
    "/login": {
      "post": {
        "tags": [
          "login-ctrl"
        ],
        "summary": "로그인 api",
        "description": "로그인 api",
        "operationId": "prevLogin",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequestInfo"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/login/checkSmsNum": {
      "post": {
        "tags": [
          "login-ctrl"
        ],
        "summary": "sms 인증번호 확인 - 비밀번호 초기화 전용",
        "description": "로그인 필요 없음",
        "operationId": "checkSmsNumWithoutSession",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId": {
                    "type": "string",
                    "example": "admin"
                  },
                  "smsNum": {
                    "type": "string",
                    "example": 123456
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/login/check/sendSmsToUser": {
      "post": {
        "tags": [
          "login-ctrl"
        ],
        "summary": "sms 발송요청",
        "description": "sms 발송요청",
        "operationId": "sendSmsToUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId": {
                    "type": "string",
                    "example": "admin"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/login/check/resetPwd": {
      "post": {
        "tags": [
          "login-ctrl"
        ],
        "summary": "비밀번호 초기화",
        "description": "비밀번호 초기화",
        "operationId": "resetPwd",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId": {
                    "type": "string",
                    "example": "admin"
                  },
                  "newPwd": {
                    "type": "string",
                    "example": "roqkf00!"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/login/check/checkSmsNum": {
      "post": {
        "tags": [
          "login-ctrl"
        ],
        "summary": "sms 인증",
        "description": "로그인 후 sms 인증",
        "operationId": "checkSmsNum",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "smsNum": {
                    "type": "string",
                    "example": 123456
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/firmware/retrieveFirmwareStatusList": {
      "post": {
        "tags": [
          "디바이스 관리 - 펌웨어 상태"
        ],
        "summary": "펌웨어 상태 정보 조회",
        "description": "펌웨어 상태 정보 조회",
        "operationId": "retrieveFirmwareStatusList",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "svcCd": {
                    "type": "string",
                    "example": "LK01"
                  },
                  "deviceCd": {
                    "type": "string",
                    "example": 1
                  },
                  "systemType": {
                    "type": "string",
                    "example": "D"
                  },
                  "fwVer": {
                    "type": "string",
                    "example": ""
                  },
                  "ctn": {
                    "type": "string",
                    "example": ""
                  },
                  "entityId": {
                    "type": "string",
                    "example": ""
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComResponseFirmwareStatusList"
                }
              }
            }
          }
        }
      }
    },
    "/firmware/downloadFirmwareStatusList": {
      "post": {
        "tags": [
          "디바이스 관리 - 펌웨어 상태"
        ],
        "summary": "펌웨어 상태 정보 엑셀 다운로드",
        "description": "펌웨어 상태 정보 엑셀 다운로드",
        "operationId": "downloadFirmwareStatusList",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "svcCd": {
                    "type": "string",
                    "example": "LK01"
                  },
                  "deviceCd": {
                    "type": "string",
                    "example": 1
                  },
                  "systemType": {
                    "type": "string",
                    "example": "D"
                  },
                  "fwVer": {
                    "type": "string",
                    "example": ""
                  },
                  "ctn": {
                    "type": "string",
                    "example": ""
                  },
                  "deviceSerialNo": {
                    "type": "string",
                    "example": ""
                  },
                  "currentPage": {
                    "type": "string",
                    "example": 1
                  },
                  "limit": {
                    "type": "string",
                    "example": 10
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "LoginRequestInfo": {
        "required": [
          "info1",
          "info2"
        ],
        "type": "object",
        "properties": {
          "info1": {
            "type": "string"
          },
          "info2": {
            "type": "string"
          }
        }
      },
...
      "ComResponseFirmwareStatusList": {
        "type": "object",
        "properties": {
          "httpStatusCode": {
            "type": "integer",
            "format": "int32"
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "detailMessage": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "body": {
            "$ref": "#/components/schemas/FirmwareStatusList"
          },
          "errorMessage": {
            "type": "string"
          }
        }
      },
      "FirmwareStatus": {
        "type": "object",
        "properties": {
          "entityId": {
            "type": "string"
          },
          "fwCd": {
            "type": "string"
          },
          "fwNm": {
            "type": "string"
          },
          "svcCd": {
            "type": "string"
          },
          "svcNm": {
            "type": "string"
          },
          "deviceCd": {
            "type": "string"
          },
          "deviceNm": {
            "type": "string"
          },
          "fwVer": {
            "type": "string"
          },
          "betaVerYn": {
            "type": "string"
          },
          "mdn": {
            "type": "string"
          },
          "updDt": {
            "type": "string"
          },
          "systemType": {
            "type": "string"
          },
          "fwAction": {
            "type": "string"
          },
          "fwStatus": {
            "type": "string"
          },
          "mac": {
            "type": "string"
          }
        }
      },
      "FirmwareStatusList": {
        "type": "object",
        "properties": {
          "list": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FirmwareStatus"
            }
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
    }
  }
}