# spring/generated-sources/openapi/src/main/java/com/openapi/generator/aicc/api/FirmwareApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.18.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.openapi.generator.aicc.api;

import com.openapi.generator.aicc.dto.ComResponseFirmwareStatusList;
import com.openapi.generator.aicc.dto.DownloadFirmwareStatusListRequest;
import com.openapi.generator.aicc.dto.RetrieveFirmwareStatusListRequest;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-01-19T16:30:08.398412+09:00[Asia/Seoul]", comments = "Generator version: 7.18.0")
@Validated
@Tag(name = "디바이스 관리 - 펌웨어 상태", description = "디바이스 관리 - 펌웨어 상태")
public interface FirmwareApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    String PATH_DOWNLOAD_FIRMWARE_STATUS_LIST = "/firmware/downloadFirmwareStatusList";
    /**
     * POST /firmware/downloadFirmwareStatusList : 펌웨어 상태 정보 엑셀 다운로드
     * 펌웨어 상태 정보 엑셀 다운로드
     *
     * @param downloadFirmwareStatusListRequest  (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "downloadFirmwareStatusList",
        summary = "펌웨어 상태 정보 엑셀 다운로드",
        description = "펌웨어 상태 정보 엑셀 다운로드",
        tags = { "디바이스 관리 - 펌웨어 상태" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = FirmwareApi.PATH_DOWNLOAD_FIRMWARE_STATUS_LIST,
        consumes = { "application/json" }
    )
    default ResponseEntity<Void> downloadFirmwareStatusList(
        @Parameter(name = "DownloadFirmwareStatusListRequest", description = "", required = true) @Valid @RequestBody DownloadFirmwareStatusListRequest downloadFirmwareStatusListRequest
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_RETRIEVE_FIRMWARE_STATUS_LIST = "/firmware/retrieveFirmwareStatusList";
    /**
     * POST /firmware/retrieveFirmwareStatusList : 펌웨어 상태 정보 조회
     * 펌웨어 상태 정보 조회
     *
     * @param retrieveFirmwareStatusListRequest  (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "retrieveFirmwareStatusList",
        summary = "펌웨어 상태 정보 조회",
        description = "펌웨어 상태 정보 조회",
        tags = { "디바이스 관리 - 펌웨어 상태" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ComResponseFirmwareStatusList.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = FirmwareApi.PATH_RETRIEVE_FIRMWARE_STATUS_LIST,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<ComResponseFirmwareStatusList> retrieveFirmwareStatusList(
        @Parameter(name = "RetrieveFirmwareStatusListRequest", description = "", required = true) @Valid @RequestBody RetrieveFirmwareStatusListRequest retrieveFirmwareStatusListRequest
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : \"code\", \"errorMessage\" : \"errorMessage\", \"detailMessage\" : \"detailMessage\", \"message\" : \"message\", \"body\" : { \"list\" : [ { \"fwNm\" : \"fwNm\", \"fwCd\" : \"fwCd\", \"betaVerYn\" : \"betaVerYn\", \"fwVer\" : \"fwVer\", \"deviceNm\" : \"deviceNm\", \"entityId\" : \"entityId\", \"svcCd\" : \"svcCd\", \"deviceCd\" : \"deviceCd\", \"mac\" : \"mac\", \"fwStatus\" : \"fwStatus\", \"svcNm\" : \"svcNm\", \"mdn\" : \"mdn\", \"fwAction\" : \"fwAction\", \"systemType\" : \"systemType\", \"updDt\" : \"updDt\" }, { \"fwNm\" : \"fwNm\", \"fwCd\" : \"fwCd\", \"betaVerYn\" : \"betaVerYn\", \"fwVer\" : \"fwVer\", \"deviceNm\" : \"deviceNm\", \"entityId\" : \"entityId\", \"svcCd\" : \"svcCd\", \"deviceCd\" : \"deviceCd\", \"mac\" : \"mac\", \"fwStatus\" : \"fwStatus\", \"svcNm\" : \"svcNm\", \"mdn\" : \"mdn\", \"fwAction\" : \"fwAction\", \"systemType\" : \"systemType\", \"updDt\" : \"updDt\" } ], \"totalCount\" : 6 }, \"httpStatusCode\" : 0, \"status\" : \"status\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}