# spring/generated-sources/openapi/src/main/java/com/openapi/generator/aicc/api/LoginApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.18.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.openapi.generator.aicc.api;

import com.openapi.generator.aicc.dto.CheckSmsNumRequest;
import com.openapi.generator.aicc.dto.CheckSmsNumWithoutSessionRequest;
import com.openapi.generator.aicc.dto.LoginRequestInfo;
import com.openapi.generator.aicc.dto.ResetPwdRequest;
import com.openapi.generator.aicc.dto.SendSmsToUserRequest;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-01-19T16:30:08.398412+09:00[Asia/Seoul]", comments = "Generator version: 7.18.0")
@Validated
@Tag(name = "login-ctrl", description = "the login-ctrl API")
public interface LoginApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    String PATH_CHECK_SMS_NUM = "/login/check/checkSmsNum";
    /**
     * POST /login/check/checkSmsNum : sms 인증
     * 로그인 후 sms 인증
     *
     * @param checkSmsNumRequest  (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "checkSmsNum",
        summary = "sms 인증",
        description = "로그인 후 sms 인증",
        tags = { "login-ctrl" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = LoginApi.PATH_CHECK_SMS_NUM,
        consumes = { "application/json" }
    )
    default ResponseEntity<Void> checkSmsNum(
        @Parameter(name = "CheckSmsNumRequest", description = "", required = true) @Valid @RequestBody CheckSmsNumRequest checkSmsNumRequest
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_CHECK_SMS_NUM_WITHOUT_SESSION = "/login/checkSmsNum";
    /**
     * POST /login/checkSmsNum : sms 인증번호 확인 - 비밀번호 초기화 전용
     * 로그인 필요 없음
     *
     * @param checkSmsNumWithoutSessionRequest  (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "checkSmsNumWithoutSession",
        summary = "sms 인증번호 확인 - 비밀번호 초기화 전용",
        description = "로그인 필요 없음",
        tags = { "login-ctrl" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = LoginApi.PATH_CHECK_SMS_NUM_WITHOUT_SESSION,
        consumes = { "application/json" }
    )
    default ResponseEntity<Void> checkSmsNumWithoutSession(
        @Parameter(name = "CheckSmsNumWithoutSessionRequest", description = "", required = true) @Valid @RequestBody CheckSmsNumWithoutSessionRequest checkSmsNumWithoutSessionRequest
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_PREV_LOGIN = "/login";
    /**
     * POST /login : 로그인 api
     * 로그인 api
     *
     * @param loginRequestInfo  (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "prevLogin",
        summary = "로그인 api",
        description = "로그인 api",
        tags = { "login-ctrl" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = LoginApi.PATH_PREV_LOGIN,
        consumes = { "application/json" }
    )
    default ResponseEntity<Void> prevLogin(
        @Parameter(name = "LoginRequestInfo", description = "", required = true) @Valid @RequestBody LoginRequestInfo loginRequestInfo
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_RESEND_SMS = "/login/resendSms";
    /**
     * POST /login/resendSms : sms 인증번호 재발송
     * sms 인증번호 재발송 (3회 제한)
     *
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "resendSms",
        summary = "sms 인증번호 재발송",
        description = "sms 인증번호 재발송 (3회 제한)",
        tags = { "login-ctrl" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = LoginApi.PATH_RESEND_SMS
    )
    default ResponseEntity<Void> resendSms(
        
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_RESET_PWD = "/login/check/resetPwd";
    /**
     * POST /login/check/resetPwd : 비밀번호 초기화
     * 비밀번호 초기화
     *
     * @param resetPwdRequest  (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "resetPwd",
        summary = "비밀번호 초기화",
        description = "비밀번호 초기화",
        tags = { "login-ctrl" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = LoginApi.PATH_RESET_PWD,
        consumes = { "application/json" }
    )
    default ResponseEntity<Void> resetPwd(
        @Parameter(name = "ResetPwdRequest", description = "", required = true) @Valid @RequestBody ResetPwdRequest resetPwdRequest
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_SEND_SMS_TO_USER = "/login/check/sendSmsToUser";
    /**
     * POST /login/check/sendSmsToUser : sms 발송요청
     * sms 발송요청
     *
     * @param sendSmsToUserRequest  (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "sendSmsToUser",
        summary = "sms 발송요청",
        description = "sms 발송요청",
        tags = { "login-ctrl" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = LoginApi.PATH_SEND_SMS_TO_USER,
        consumes = { "application/json" }
    )
    default ResponseEntity<Void> sendSmsToUser(
        @Parameter(name = "SendSmsToUserRequest", description = "", required = true) @Valid @RequestBody SendSmsToUserRequest sendSmsToUserRequest
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}